+++ frankensteinsmonster/CodeFile1.cs
@@ -53,8 +53,8 @@ namespace Rextester // this is a hold over from online testing
         public int Fast_exp(int base_num, int pow, int mod)
         {
             // uses repeated squaring method
-            int tracker = base_num % mod;
-            int output = 1;
+            long tracker = base_num % mod;
+            long output = 1;
             while (pow > 0)
             {
                 if (pow % 2 != 0)
@@ -64,7 +64,7 @@ namespace Rextester // this is a hold over from online testing
                 tracker = (tracker * tracker) % mod;
                 pow = pow / 2;
             }
-            return output;
+            return (int)output;
         }
 
         public string Sanitize_int(string message)
